Revert "Ensure hydration scripts inside of slots render ASAP (#4288)"
This reverts commit c218100684c90c2b5c490e73b0687ad59d0c58df.
