+++ packages/astro/src/runtime/server/render/astro.ts
@@ -1,5 +1,6 @@
-import type { SSRRenderInstruction, SSRResult } from '../../../@types/astro';
+import type { SSRResult } from '../../../@types/astro';
 import type { AstroComponentFactory } from './index';
+import type { RenderInstruction } from './types';
 
 import { markHTMLString } from '../escape.js';
 import { HydrationDirectiveProps } from '../hydration.js';
@@ -57,7 +58,7 @@ export function isAstroComponent(obj: any): obj is AstroComponent {
 
 export async function* renderAstroComponent(
 	component: InstanceType<typeof AstroComponent>
-): AsyncIterable<string | SSRRenderInstruction> {
+): AsyncIterable<string | RenderInstruction> {
 	for await (const value of component) {
 		if (value || value === 0) {
 			for await (const chunk of renderChild(value)) {
@@ -103,7 +104,7 @@ export async function renderToIterable(
 	displayName: string,
 	props: any,
 	children: any
-): Promise<AsyncIterable<string | SSRRenderInstruction>> {
+): Promise<AsyncIterable<string | RenderInstruction>> {
 	validateComponentProps(props, displayName);
 	const Component = await componentFactory(result, props, children);
