+++ packages/astro/src/runtime/server/render/component.ts
@@ -1,9 +1,8 @@
 import type {
 	AstroComponentMetadata,
-	SSRLoadedRenderer,
-	SSRRenderInstruction,
-	SSRResult,
+	SSRLoadedRenderer, SSRResult
 } from '../../../@types/astro';
+import type { RenderInstruction } from './types.js';
 
 import { markHTMLString } from '../escape.js';
 import { extractDirectives, generateHydrateScript } from '../hydration.js';
@@ -53,7 +52,7 @@ export async function renderComponent(
 	Component: unknown,
 	_props: Record<string | number, any>,
 	slots: any = {}
-): Promise<string | AsyncIterable<string | SSRRenderInstruction>> {
+): Promise<string | AsyncIterable<string | RenderInstruction>> {
 	Component = await Component;
 
 	switch (getComponentType(Component)) {
@@ -83,7 +82,7 @@ export async function renderComponent(
 
 		case 'astro-factory': {
 			async function* renderAstroComponentInline(): AsyncGenerator<
-				string | SSRRenderInstruction,
+				string | RenderInstruction,
 				void,
 				undefined
 			> {
