+++ packages/astro/src/runtime/server/render/types.ts
@@ -1,0 +1,8 @@
+import type { SSRResult } from '../../../@types/astro';
+import type { HydrationMetadata } from '../hydration.js';
+
+export interface RenderInstruction {
+	type: 'directive';
+	result: SSRResult;
+	hydration: HydrationMetadata;
+}
