--- packages/astro/test/fixtures/hydration-race/src/components/Layout.astro
@@ -1,6 +1,0 @@
----
-import Wrapper from './Wrapper.astro';
----
-
-<Wrapper />
-<slot />
