--- packages/astro/test/fixtures/hydration-race/src/pages/slot.astro
@@ -1,15 +1,0 @@
----
-import Layout from '../components/Layout.astro';
-import Wrapper from '../components/Wrapper.astro';
-import One from '../components/One.jsx';
----
-<html>
-<head>
-	<title>Testing</title>
-</head>
-<body>
-	<Layout>
-		<One name="Three" client:idle />
-	</Layout>
-</body>
-</html>
